+++ src/main.ts
@@ -45,7 +45,7 @@ b_reset.addEventListener("click",e=>{
 });
 
 function setPlayI(i:number){
-    let str = localStorage.getItem("__SE-PD") || "{locId:0}";
+    let str = localStorage.getItem("__SE-PD") || '{"locId":0}';
     let o = JSON.parse(str);
     o.locId = i;
     localStorage.setItem("__SE-PD",JSON.stringify(o));